requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pdfminer.six==20231228
docx2txt==0.8
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
nltk==3.8.1
